\name{risaac-package}
\alias{risaac-package}
\docType{package}
\title{A fast cryptographic random number generator using ISAAC by Robert Jenkins.}
\description{Uses the ISAAC generator: http://burtleburtle.net/bob/rand/isaacafa.html}
\details{
\tabular{ll}{Package: \tab risaac\cr
Type: \tab Package\cr
Title: \tab A fast cryptographic random number generator using ISAAC by Robert Jenkins.\cr
Version: \tab 0.1\cr
Date: \tab 2011-08-23\cr
Author: \tab David L Gibbs\cr
Maintainer: \tab David L Gibbs <gibbsdavidl@gmail.com>\cr
License: \tab GPL-v3\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2010-09-20 22:44:55 UTC; dave\cr}
}
\author{David L Gibbs}

\keyword{ package }


